web/requirements.dev.txt
-r requirements.common.txt

pytest==7.1.3
pytest-cov==4.0.0
pytest-django==4.5.2
pre-commit==2.10.1
